- findOne and findAll inventory-summary

+++ app/service/company/company.service.js
@@ -1,5 +1,5 @@
 import db from '../../db/models';
-import { badRequest, FIELD_ERROR, HTTP_ERROR, HttpError } from '../../config/error';
+import { badRequest, FIELD_ERROR } from '../../config/error';
 
 const {Op} = db.Sequelize;
 
@@ -50,7 +50,7 @@ export function createCompany(userId, createForm) {
 export async function updateCompany(cId, updateForm) {
   const company = await db.Company.findByPk(cId);
   if (!company) {
-    throw new HttpError(HTTP_ERROR.NOT_FOUND, 'Invalid Company');
+    throw badRequest('company', FIELD_ERROR.INVALID, 'company not found');
   }
   company.name = updateForm.name;
   company.gsm = updateForm.gsm;
@@ -59,8 +59,12 @@ export async function updateCompany(cId, updateForm) {
   return company.save();
 }
 
-export function removeCompany (cId) {
+export async function removeCompany (cId) {
+  const company = await db.Company.findByPk(cId);
+  if (!company) {
+    throw badRequest('company', FIELD_ERROR.INVALID, 'company not found');
+  }
   return db.Company.destroy({
-    where: { id: cId }
+    where: { id: company.id }
   });
 }
